week5/requirements.txt
langchain
chromadb
openai
python-dotenv
requests